Moved testcase outside fixture folder
